single precision changed to double precision

+++ lib/transform_data_structure.py
@@ -6,9 +6,9 @@ import numpy as np
 def reorganize_by_prim_elec(waveformsL, times, spike_idsL, prim_elecL):
     n_channels = max(prim_elecL) + 1
     n_samples = waveformsL[0].shape[1]
-    wvf_byc = [np.zeros((0, n_samples), dtype=np.single) for _ in range(n_channels)]
+    wvf_byc = [np.zeros((0, n_samples), dtype=np.double) for _ in range(n_channels)]
     # setting the elements of times_byc and units_byc to be in the 2-D array shape that will be converted to Matlab matrices
-    times_byc = [np.zeros((0), dtype=np.single)[np.newaxis].transpose() for _ in range(n_channels)]
+    times_byc = [np.zeros((0), dtype=np.double)[np.newaxis].transpose() for _ in range(n_channels)]
     units_byc = [np.zeros((0), dtype=np.int16)[np.newaxis].transpose() for _ in range(n_channels)]
     for cluster_id, (waveforms, spike_ids, prim_elec) in enumerate(zip(waveformsL, spike_idsL, prim_elecL)):        
         # print(cluster_id, "", waveforms.shape, "", prim_elec)
